Learn DOM Manipulation and Events java
Created a new directory for the lesson.
Created new index.html, java.js, and style.css that will be used in
the lesson.
Inside the files are the things I learned like, creating new elements
for the html file and styling them, from the java file.
Refreshed my learnings in linking the java and css files to the html
files.
Started learning about events.